refactor(db): consolidate the three duplicate D1 read-model schema mirrors into a @kampus/db-schema leaf
The worker, preview-seed, and fts-backfill each hand-maintained a Drizzle
mirror of the same D1 read-model tables (term_summary / definition_view /
post_summary / comment_view), pinned only by a 'mirror the canonical
migration' docblock. A column change drifted silently — ADR 0093's is_draft
never reached preview-seed, and ADR 0096's deleted_at→removed_at broke both
copies — caught only by real-D1 CI at deploy/runtime, never by typecheck.

Introduce @kampus/db-schema: a leaf package (depends only on drizzle-orm)
holding the single canonical declaration of the four shared read-model tables.
The worker's schema.ts re-exports them (its 15 import sites unchanged);
preview-seed and fts-backfill re-export from it too. A column rename is now ONE
edit, reflected everywhere by construction and caught by typecheck.

Leaf, not direct worker-schema import: fts-backfill already prod-depends on
@kampus/web and the repo deliberately keeps apps/web→fts-backfill off the
graph, so the shared source must be a leaf both depend on (no cycle) — this is
#903's load-bearing constraint.

Also closes #859's tactical drift items: preview-seed's test DDL gains the
removed_by/removed_reason/is_draft columns the canonical tables carry; the
'three tables' prose is corrected to five (3 read-model + 2 FTS5); SeedReport
surfaces the FTS row counts.

The FTS5 virtual tables stay out of the leaf (the worker never models them as
drizzle tables — they're raw-sql-synced and drizzle-kit reads the worker schema
for migration generation); preview-seed keeps its local plain-table model.

Fixes #859. Consolidates the #903 duplicate (closed as dup of #859).
