+++ apps/web/worker/db/drizzle/schema.ts
@@ -8,6 +8,15 @@
 import {sql} from "drizzle-orm";
 import {index, integer, primaryKey, sqliteTable, text} from "drizzle-orm/sqlite-core";
 
+// The shared read-model tables (`term_summary` / `definition_view` /
+// `post_summary` / `comment_view`) live in the `@kampus/db-schema` leaf so the
+// worker, preview-seed, and fts-backfill all import ONE declaration — a column
+// rename is one edit there, caught by typecheck, not three hand-mirrored copies
+// that drift (ADR 0096 `removed_at`, ADR 0093 `is_draft`; issues #859/#903).
+// drizzle-kit reads this module for migration generation, so re-exporting feeds
+// migrations exactly as a local declaration would.
+export {commentView, definitionView, postSummary, termSummary} from "@kampus/db-schema";
+
 const timestamp = (name: string) => integer(name, {mode: "timestamp"});
 
 /**
@@ -115,72 +124,6 @@ export const apikey = sqliteTable("apiKey", {
 });
 
 /**
- * One row per term, keyed by slug. Maintained by the `TermChanged` projection step.
- */
-export const termSummary = sqliteTable(
-	"term_summary",
-	{
-		slug: text("slug").primaryKey(),
-		title: text("title").notNull(),
-		// Lower-cased first character; powers the alphabet pivot on SozlukHome.
-		firstLetter: text("first_letter").notNull(),
-		definitionCount: integer("definition_count").notNull().default(0),
-		totalScore: integer("total_score").notNull().default(0),
-		// Body of the highest-scoring non-deleted definition, truncated.
-		excerpt: text("excerpt"),
-		topDefinitionId: text("top_definition_id"),
-		firstAt: timestamp("first_at"),
-		lastActivityAt: timestamp("last_activity_at"),
-		lastEditAt: timestamp("last_edit_at"),
-		// Convergent-overwrite guard: `WHERE last_event_id < excluded.last_event_id`.
-		lastEventId: text("last_event_id").notNull().default(""),
-	},
-	(t) => [
-		index("term_summary_recent").on(t.lastActivityAt),
-		index("term_summary_popular").on(t.totalScore),
-		index("term_summary_letter").on(t.firstLetter),
-	],
-);
-
-/**
- * Per-definition row. Canonical store for sozluk definitions after d1-direct
- * (ADR 0009) — the per-term DO is no longer the source of truth. Denormalized
- * with term slug + title so the profile feed renders without joining
- * `term_summary`; `body_excerpt` is a denormalized truncation for the feed card.
- *
- * `last_event_id` is vestigial (projection-era convergence guard, unused under
- * d1-direct); kept to hold the read-side schema stable until a cleanup pass.
- */
-export const definitionView = sqliteTable(
-	"definition_view",
-	{
-		id: text("id").primaryKey(),
-		authorId: text("author_id").notNull(),
-		authorName: text("author_name").notNull(),
-		termSlug: text("term_slug").notNull(),
-		termTitle: text("term_title").notNull(),
-		body: text("body").notNull().default(""),
-		bodyExcerpt: text("body_excerpt").notNull(),
-		score: integer("score").notNull().default(0),
-		createdAt: timestamp("created_at").notNull(),
-		updatedAt: timestamp("updated_at").notNull(),
-		// The ADR 0096 removal triad. `removed_at` null ⇒ Live; this column IS the
-		// former `deleted_at`, repurposed. `removed_by`/`removed_reason` carry the
-		// audit. Projected to `EntityLifecycle` — services never read these raw.
-		removedAt: timestamp("removed_at"),
-		removedBy: text("removed_by"),
-		removedReason: text("removed_reason"),
-		lastEventId: text("last_event_id").notNull().default(""),
-	},
-	(t) => [
-		// WHERE author_id = ? ORDER BY created_at DESC (profile feed).
-		index("definition_view_author_created").on(t.authorId, t.createdAt),
-		// WHERE term_slug = ? AND removed_at IS NULL (term page).
-		index("definition_view_term_score").on(t.termSlug, t.score),
-	],
-);
-
-/**
  * Per-(definition, voter) up-vote presence row. `user_vote` and
  * `definition_view.score` (COUNT(*) under `WHERE definition_id = ?`) are both
  * denormalized off this, recomputed inline in the same D1 batch as the vote write.
@@ -210,60 +153,6 @@ export const sozlukStats = sqliteTable("sozluk_stats", {
 	updatedAt: timestamp("updated_at").notNull(),
 });
 
-/**
- * One row per post. Maintained by the `PostChanged` projection step.
- */
-export const postSummary = sqliteTable(
-	"post_summary",
-	{
-		id: text("id").primaryKey(),
-		slug: text("slug"),
-		title: text("title").notNull(),
-		// Submission URL (link posts). Denormalized so the feed renders without
-		// RPCing into the per-post DO.
-		url: text("url"),
-		// Extracted via `new URL(url).host` on submit. Powers the host filter.
-		host: text("host"),
-		// Canonical full-text body under D1-direct; `body_excerpt` stays for feed cards.
-		body: text("body").notNull().default(""),
-		bodyExcerpt: text("body_excerpt"),
-		authorId: text("author_id").notNull(),
-		authorName: text("author_name").notNull(),
-		// Comma-separated tags from the fixed enum (göster/tartışma/soru/söylenme/meta).
-		tags: text("tags").notNull().default(""),
-		score: integer("score").notNull().default(0),
-		commentCount: integer("comment_count").notNull().default(0),
-		// HN-style hot score: f(score, age). Recomputed on every PostChanged.
-		hotScore: integer("hot_score").notNull().default(0),
-		createdAt: timestamp("created_at").notNull(),
-		updatedAt: timestamp("updated_at").notNull(),
-		lastActivityAt: timestamp("last_activity_at").notNull(),
-		// The ADR 0096 removal triad (see `definition_view`). `removed_at` is the
-		// former `deleted_at`. Pano posts that hard-deleted pre-substrate are gone
-		// and not reconstructable; new removals are soft `Removed`, karma kept.
-		removedAt: timestamp("removed_at"),
-		removedBy: text("removed_by"),
-		removedReason: text("removed_reason"),
-		// Draft (taslak) marker — nullable, no default, mirroring the `removedAt`
-		// soft-state shape: existing/published rows are `null` (= not a draft). A
-		// partial unique index (`post_summary_one_draft_per_author`, migration 0004)
-		// enforces one draft per author. Drafts are excluded from public feeds.
-		isDraft: integer("is_draft", {mode: "boolean"}),
-		lastEventId: text("last_event_id").notNull().default(""),
-	},
-	(t) => [
-		index("post_summary_hot").on(t.hotScore),
-		index("post_summary_new").on(t.createdAt),
-		index("post_summary_top").on(t.score),
-		index("post_summary_discuss").on(t.commentCount),
-		index("post_summary_host").on(t.host),
-		// `created_at DESC` via a `sql` fragment: drizzle 0.45's index DSL can't
-		// express per-column ordering, and SQLite must walk forward for the
-		// newest-first profile feed read.
-		index("post_summary_author_created").on(t.authorId, sql`${t.createdAt} DESC`),
-	],
-);
-
 /** Per-(post, voter) up-vote presence row. Mirrors `definitionVote`. */
 export const postVote = sqliteTable(
 	"post_vote",
@@ -292,52 +181,6 @@ export const postBookmark = sqliteTable(
 	(t) => [
 		primaryKey({columns: [t.postId, t.userId]}),
 		index("post_bookmark_user_created").on(t.userId, sql`${t.createdAt} DESC`),
-	],
-);
-
-/**
- * Per-comment row, denormalized with post id + title for the profile feed AND
- * the per-post thread reader. Canonical store for pano comments after d1-direct
- * (ADR 0009) — the per-post DO is no longer the source of truth.
- *
- * A removed comment that still has live replies stays as a `Removed` row (its
- * canonical body kept for restore + moderator review); the `[silindi]` tombstone
- * is a VIEW rendering of that state, not a body the delete path writes (ADR 0096
- * §5). A removed leaf comment is also a `Removed` row now — no hard delete.
- *
- * `last_event_id` is vestigial (projection-era convergence guard, unused under
- * d1-direct); kept to hold the read-side schema stable until a cleanup pass.
- */
-export const commentView = sqliteTable(
-	"comment_view",
-	{
-		id: text("id").primaryKey(),
-		authorId: text("author_id").notNull(),
-		authorName: text("author_name").notNull(),
-		postId: text("post_id").notNull(),
-		postTitle: text("post_title").notNull(),
-		// NULL for top-level comments; nested replies point at a non-removed
-		// comment in the same post.
-		parentId: text("parent_id"),
-		body: text("body").notNull().default(""),
-		bodyExcerpt: text("body_excerpt").notNull(),
-		score: integer("score").notNull().default(0),
-		createdAt: timestamp("created_at").notNull(),
-		updatedAt: timestamp("updated_at").notNull(),
-		// The ADR 0096 removal triad (see `definition_view`). `removed_at` is the
-		// former `deleted_at`.
-		removedAt: timestamp("removed_at"),
-		removedBy: text("removed_by"),
-		removedReason: text("removed_reason"),
-		lastEventId: text("last_event_id").notNull().default(""),
-	},
-	(t) => [
-		// WHERE author_id = ? ORDER BY created_at DESC (profile feed).
-		index("comment_view_author_created").on(t.authorId, t.createdAt),
-		// WHERE post_id = ? ORDER BY created_at ASC (per-post thread).
-		index("comment_view_post").on(t.postId),
-		// Reply-aware soft-delete children-of-parent check.
-		index("comment_view_parent").on(t.parentId),
 	],
 );
