+++ packages/db-schema/package.json
@@ -1,0 +1,27 @@
+{
+	"name": "@kampus/db-schema",
+	"version": "0.0.0",
+	"private": true,
+	"description": "Single canonical Drizzle declaration of the shared D1 read-model tables (term_summary / definition_view / post_summary / comment_view) — the one source the worker, preview-seed, and fts-backfill all import so a column rename is one edit (issue #859)",
+	"type": "module",
+	"exports": {
+		".": "./src/index.ts"
+	},
+	"scripts": {
+		"typecheck": "tsgo -p tsconfig.json",
+		"test": "vitest run"
+	},
+	"dependencies": {
+		"drizzle-orm": "catalog:"
+	},
+	"devDependencies": {
+		"@effect/tsgo": "catalog:",
+		"@types/node": "catalog:",
+		"@typescript/native-preview": "catalog:",
+		"typescript": "catalog:",
+		"vitest": "catalog:"
+	},
+	"volta": {
+		"node": "26.2.0"
+	}
+}
